rename: 클라이언트 코드를 main 함수에서 Controller로 이동
